Cleaning up Interval Checks (#163)
* draft of inteval check edits

* Fixing up quadratic checks

* updated unit tests, fixed small bug
